fix: metrics auto-refresh when a query is executed (#42)
This pull request includes changes to the `CqlEditor` component to
improve its rendering logic, specifically by forcing re-renders when the
query is executed. The key updates involve introducing a new state
variable and utilizing the `useCallback` hook for rendering.

Rendering improvements:

*
[`src/app/(main)/query-runner/_components/cql-editor.tsx`](diffhunk://#diff-c3d008cee84b65d82023278c89d53a325b14f0df0e41dfe05a6b97bc57e63f16R52):
Added a new state variable `updateKey` to force re-render when the query
is executed.
*
[`src/app/(main)/query-runner/_components/cql-editor.tsx`](diffhunk://#diff-c3d008cee84b65d82023278c89d53a325b14f0df0e41dfe05a6b97bc57e63f16R68):
Incremented `updateKey` in the `onSuccess` callback to trigger a
re-render.
*
[`src/app/(main)/query-runner/_components/cql-editor.tsx`](diffhunk://#diff-c3d008cee84b65d82023278c89d53a325b14f0df0e41dfe05a6b97bc57e63f16L207-R236):
Replaced the `renderTabs` object with individual `useCallback` hooks for
`renderResult`, `renderTracing`, and `renderDashboard` to ensure
components re-render correctly with the updated `updateKey`.